requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
distlib==0.3.1
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
filelock==3.0.12
pipenv==2020.6.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.27
virtualenv-clone==0.5.4
